+++ src/meico/mei/Mei.java
@@ -2,6 +2,7 @@ package meico.mei;
 
 import meico.mpm.Mpm;
 import meico.msm.Msm;
+import meico.musicxml.MusicXml;
 import meico.supplementary.KeyValue;
 import meico.svg.SvgCollection;
 import nu.xom.*;
@@ -294,7 +295,7 @@ public class Mei extends meico.xml.XmlBase {
     /**
      * converts the mei data into msm and mpm format and returns a tuplet of lists, one with the msms (one per movement/mdiv), the other with the corresponding mpms
      * @param ppq the ppq resolution for the conversion; this is counterchecked with the minimal required resolution to capture the shortest duration in the mei data; if a higher resolution is necessary, this input parameter is overridden
-     * @param dontUseChannel10 the flag says whether channel 10 (midi drum channel) shall be used or not; it is already dont here, at the mei2msm conversion, because the msm should align with the midi file later on
+     * @param dontUseChannel10 the flag says whether channel 10 (midi drum channel) shall be used or not; it is already done here, at the mei2msm conversion, because the msm should align with the midi file later on
      * @return the list of msm documents (movements) created
      */
     public KeyValue<List<Msm>, List<Mpm>> exportMsmMpm(int ppq, boolean dontUseChannel10) {
@@ -304,7 +305,7 @@ public class Mei extends meico.xml.XmlBase {
     /**
      * converts the mei data into msm and mpm format and returns a tuplet of lists, one with the msms (one per movement/mdiv), the other with the corresponding mpms
      * @param ppq the ppq resolution for the conversion; this is counterchecked with the minimal required resolution to capture the shortest duration in the mei data; if a higher resolution is necessary, this input parameter is overridden
-     * @param dontUseChannel10 the flag says whether channel 10 (midi drum channel) shall be used or not; it is already dont here, at the mei2msm conversion, because the msm should align with the midi file later on
+     * @param dontUseChannel10 the flag says whether channel 10 (midi drum channel) shall be used or not; it is already done here, at the mei2msm conversion, because the msm should align with the midi file later on
      * @param ignoreExpansions set this true to have a 1:1 conversion of MEI to MSM without the rearrangement that MEI's expansion elements produce
      * @return the list of msm documents (movements) created
      */
@@ -316,12 +317,30 @@ public class Mei extends meico.xml.XmlBase {
      * converts the mei data into msm and mpm format and returns a tuplet of lists, one with the msms (one per movement/mdiv), the other with the corresponding mpms
      * @param ppq the ppq resolution for the conversion; this is counterchecked with the minimal required resolution to capture the shortest duration in the mei data; if a higher resolution is necessary, this input parameter is overridden
      * @param dontUseChannel10 the flag says whether channel 10 (midi drum channel) shall be used or not; it is already done here, at the mei2msm conversion, because the msm should align with the midi file later on
-     * @param cleanup set true to return a clean msm file or false to keep all the crap from the conversion
      * @param ignoreExpansions set this true to have a 1:1 conversion of MEI to MSM without the rearrangement that MEI's expansion elements produce
+     * @param cleanup set true to return a clean msm file or false to keep all the crap from the conversion
      * @return the list of msm documents (movements) created
      */
     public synchronized KeyValue<List<Msm>, List<Mpm>> exportMsmMpm(int ppq, boolean dontUseChannel10, boolean ignoreExpansions, boolean cleanup) {
         return (new Mei2MsmMpmConverter(ppq, dontUseChannel10, ignoreExpansions, cleanup)).convert(this);
+    }
+
+    /**
+     * converts the MEI data to MusicXML and returns a list of MusicXml objects, one per movement/mdiv;
+     * MEI expansions are resolved
+     * @return
+     */
+    public synchronized List<MusicXml> exportMusicXml() {
+        return this.exportMusicXml(false);
+    }
+
+    /**
+     * converts the MEI data to MusicXML and returns a list of MusicXml objects, one per movement/mdiv
+     * @param ignoreExpansions set this true to have a 1:1 conversion of MEI to MusicXML without the rearrangement that MEI's expansion elements produce
+     * @return
+     */
+    public synchronized List<MusicXml> exportMusicXml(boolean ignoreExpansions) {
+        return (new Mei2MusicXmlConverter(ignoreExpansions)).convert(this);
     }
 
     /**